ai_phising/requirements.txt
cryptography
boto3
google-generativeai
